Experimente

+++ gui/include/DummyData.hpp
@@ -24,7 +24,7 @@ struct Vertex_XYZ_RGB_ST
 {
     GLfloat position[3];
     GLubyte color[3];
-    GLfloat texture[2];
+    GLshort texture[2];
 };
 
 struct Vertex_XYZ_NRM_RGB
@@ -38,7 +38,7 @@ struct Vertex_XYZ_NRM_RGB_ST
 {
     GLfloat position[3];
     GLfloat normal[3];
-    GLfloat texture[2];
+    GLushort texture[2];
     GLubyte color[3];
 };
 #pragma pack(pop)
